--- MyList.rb
@@ -1,12 +1,0 @@
-require_relative 'MyEnumerable'
-
-class MyList 
-  include MyEnumerable
-    def initialize(list)
-      @list = list 
-    end
-
-    def each(&block)
-      @list.each(&block)
-    end
-end
