+++ my_list.rb
@@ -1,0 +1,12 @@
+require_relative 'MyEnumerable'
+
+class MyList
+  include MyEnumerable
+  def initialize(list)
+    @list = list
+  end
+
+  def each(&block)
+    @list.each(&block)
+  end
+end
